Ai-recommender-agent-main/backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI and ML
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
euriai>=0.1.0
transformers==4.37.2
torch>=2.2.0
sentence-transformers==2.2.2
huggingface-hub==0.20.1

# Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
scrapy==2.11.0
requests==2.31.0
feedparser==6.0.11

# Background Tasks
celery==5.3.6
redis==5.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0
python-dateutil==2.8.2
pypdf2==3.0.1
docx2python==2.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0





